Refactor util library.
Move utils away from the __init__.py file
